@@ -1,12 +1,20 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.3.61'
-    ext.glide_version = '4.10.0'
-    ext.retrofit_version = '2.6.2'
-    ext.logging_interceptor_version = "3.10.0"
-    ext.recyclerview_version = '1.1.0'
 
+    ext {
+        kotlin_version = '1.4.32'
+        glide_version = '4.11.0'
+        recyclerview_version = '1.2.0'
+        appCompat_version = '1.2.0'
+        core_version = '1.3.2'
+        constraintLayout_version = '2.0.4'
+        jUnit_version = '4.13.1'
+        jUnitExt_version = '1.1.2'
+        espresso_version = '3.3.0'
+        androidTest_version = '1.3.0'
+        material_version = '1.3.0'
+    }
 
     repositories {
         google()
@@ -14,7 +22,7 @@ buildscript {
         
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.5.3'
+        classpath 'com.android.tools.build:gradle:4.1.3'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
